feat: Create add pool member function to admin menu options
Changed Member.kt class to accept nullable uid and email values. Created createNewPoolMember function in HomeViewModel.kt to add new member to pool after verifying required fields are valid. Created a layout resource for use in the alert dialog and connected onClickListeners in the HomeFragment.kt. Added a addNewMemberToPool function to FirestoreUtil.kt that checks if member already exists before adding it to the pool.